+++ bin/updater/src/run_updater.sh
