+++ bin/updater/src/update_radio_gui.py
@@ -20,7 +20,13 @@ import shutil
 import tempfile
 import zipfile
 import threading
+import subprocess
 import re
+import ssl
+import traceback
+import math
+import json
+import webbrowser
 from pathlib import Path
 from urllib.request import urlopen, Request
 from urllib.error import URLError, HTTPError
@@ -60,6 +66,16 @@ GITHUB_API_URL = "https://api.github.com/repos/rotorflight/rotorflight-lua-ethos
 ETHOS_VID = 0x0483
 ETHOS_PID = 0x5750
 TARGET_NAME = "rfsuite"
+DEFAULT_LOCALE = "en"
+AVAILABLE_LOCALES = ["en", "de", "es", "fr", "it", "nl"]
+DOWNLOAD_TIMEOUT = 120
+DOWNLOAD_RETRIES = 3
+DOWNLOAD_RETRY_DELAY = 2
+COPY_SETTLE_SECONDS = 0.02
+LOGO_URL = "https://raw.githubusercontent.com/rotorflight/rotorflight-lua-ethos-suite/master/bin/updater/src/logo.png"
+UPDATER_VERSION = "0.0.0"
+UPDATER_RELEASE_JSON_URL = "https://raw.githubusercontent.com/rotorflight/rotorflight-lua-ethos-suite/master/bin/updater/src/release.json"
+UPDATER_INFO_URL = "https://github.com/rotorflight/rotorflight-lua-ethos-suite/tree/master/bin/updater/"
 
 # Version types
 VERSION_RELEASE = "release"
@@ -70,202 +86,6 @@ VERSION_MASTER = "master"
 ETHOS_SUITE_USB_MODE_REQUEST = 0x81
 USB_MODE_STORAGE = 0x69  # Stop debug mode = enable storage mode
 USB_MODE_DEBUG = 0x68    # Start debug mode
-
-TAG_RE = re.compile(
-    r'@i18n\(\s*([^)@,]+?)\s*(?:,\s*(upper|lower))?\s*\)'
-    r'((?::[a-z_]+(?:\([^@]*?\))?)*)@',
-    flags=re.IGNORECASE
-)
-
-I18N_FILE_EXTS = ('.lua', '.ts', '.tsx', '.js', '.jsx', '.json', '.md', '.txt')
-
-
-
-def _coerce_atom(value):
-    try:
-        return int(value)
-    except ValueError:
-        try:
-            return float(value)
-        except ValueError:
-            if str(value).lower() in ('true', 'false'):
-                return str(value).lower() == 'true'
-            return value
-
-
-def _parse_chain(chain):
-    if not chain:
-        return []
-    out = []
-    for seg in filter(None, chain.split(':')):
-        match = re.match(r'([a-z_][a-z0-9_]*)\s*(?:\((.*)\))?$', seg, flags=re.IGNORECASE)
-        if not match:
-            continue
-        name = match.group(1).lower()
-        argstr = (match.group(2) or '').strip()
-        args = []
-        if argstr:
-            parts = []
-            current = ''
-            depth = 0
-            for ch in argstr:
-                if ch == '(':
-                    depth += 1
-                    current += ch
-                elif ch == ')':
-                    depth = max(0, depth - 1)
-                    current += ch
-                elif ch == ',' and depth == 0:
-                    parts.append(current.strip())
-                    current = ''
-                else:
-                    current += ch
-            if current.strip():
-                parts.append(current.strip())
-            for part in parts:
-                parsed = part.strip()
-                if not parsed:
-                    args.append('')
-                elif (parsed.startswith('"') and parsed.endswith('"')) or (parsed.startswith("'") and parsed.endswith("'")):
-                    args.append(_coerce_atom(parsed[1:-1]))
-                else:
-                    args.append(_coerce_atom(parsed))
-        out.append((name, args))
-    return out
-
-
-def _upperfirst(text):
-    return text[:1].upper() + text[1:].lower() if text else text
-
-
-def _truncate(text, length, ellipsis=None):
-    length = int(length)
-    if length < 0:
-        return text
-    if len(text) <= length:
-        return text
-    if ellipsis:
-        ellipsis = str(ellipsis)
-        if length <= len(ellipsis):
-            return ellipsis[:length]
-        return text[:length - len(ellipsis)] + ellipsis
-    return text[:length]
-
-
-def _collapse_ws(text):
-    return re.sub(r'\s+', ' ', text).strip()
-
-
-def _slice(text, start, end=None):
-    start = int(start)
-    end = int(end) if end is not None else None
-    return text[start:end]
-
-
-def _ensure_char(char):
-    return char[0] if isinstance(char, str) and char else ' '
-
-
-TRANSFORMS = {
-    'upper': lambda s: s.upper(),
-    'lower': lambda s: s.lower(),
-    'upperfirst': _upperfirst,
-    'capitalize': lambda s: s[:1].upper() + s[1:],
-    'title': lambda s: s.title(),
-    'swapcase': lambda s: s.swapcase(),
-    'trim': lambda s: s.strip(),
-    'ltrim': lambda s: s.lstrip(),
-    'rtrim': lambda s: s.rstrip(),
-    'collapse_ws': _collapse_ws,
-    'truncate': lambda s, n, ellipsis=None: _truncate(s, n, ellipsis),
-    'slice': _slice,
-    'padleft': lambda s, width, char=' ': s.rjust(int(width), _ensure_char(char)),
-    'padright': lambda s, width, char=' ': s.ljust(int(width), _ensure_char(char)),
-    'center': lambda s, width, char=' ': s.center(int(width), _ensure_char(char)),
-    'replace': lambda s, old, new, count=None: s.replace(str(old), str(new), int(count) if count is not None else -1),
-    'remove': lambda s, pattern: re.sub(str(pattern), '', s),
-    'keep': lambda s, pattern: ' '.join(re.findall(str(pattern), s)),
-    'strip_prefix': lambda s, p: s[len(p):] if s.startswith(str(p)) else s,
-    'strip_suffix': lambda s, p: s[:-len(p)] if len(p) and s.endswith(str(p)) else s,
-    'prefix': lambda s, p: str(p) + s,
-    'suffix': lambda s, p: s + str(p),
-    'escape_html': lambda s: s.replace('&', '&amp;').replace('<', '&lt;').replace('>', '&gt;').replace('"', '&quot;'),
-    'escape_json': lambda s: s.replace('\\', '\\\\').replace('"', r'\"'),
-}
-
-
-def _sanitize_for_insertion(text):
-    text = text.replace("\r\n", "\n").replace("\r", "\n")
-    text = text.replace("\n", r"\n")
-    text = text.replace('"', r'\"')
-    return text
-
-
-def _resolve_key(tree, dotted):
-    node = tree
-    for part in dotted.split('.'):
-        if not isinstance(node, dict) or part not in node:
-            return None
-        node = node[part]
-
-    if isinstance(node, dict):
-        if 'translation' in node and isinstance(node['translation'], (str, int, float)):
-            return str(node['translation'])
-        if 'english' in node and isinstance(node['english'], (str, int, float)):
-            return str(node['english'])
-        return None
-
-    if node is None:
-        return None
-
-    return str(node)
-
-
-def _apply_transform_pipeline(text, basic_mod, chain, stats):
-    if basic_mod:
-        fn = TRANSFORMS.get(basic_mod.lower())
-        if fn:
-            text = fn(text)
-
-    for name, args in _parse_chain(chain):
-        fn = TRANSFORMS.get(name)
-        if not fn:
-            stats.setdefault('unknown_transform', {}).setdefault(name, 0)
-            stats['unknown_transform'][name] += 1
-            continue
-        try:
-            text = fn(text, *args)
-        except Exception as exc:
-            stats.setdefault('transform_errors', []).append(f"{name}({args}) -> {exc}")
-    return text
-
-
-def _replace_tags_in_text(text, translations, stats):
-    def _sub(match):
-        key = match.group(1).strip()
-        basic_mod = match.group(2)
-        chain = match.group(3) or ''
-
-        resolved = _resolve_key(translations, key)
-        if resolved is None:
-            stats.setdefault('unresolved', {}).setdefault(key, 0)
-            stats['unresolved'][key] += 1
-            return match.group(0)
-
-        resolved = _apply_transform_pipeline(str(resolved), basic_mod, chain, stats)
-        resolved = _sanitize_for_insertion(resolved)
-        return resolved
-
-    new_text, count = TAG_RE.subn(_sub, text)
-    return new_text, count
-
-
-def _iter_source_files(root_path, exts=I18N_FILE_EXTS):
-    for path in root_path.rglob('*'):
-        if path.is_file() and path.suffix.lower() in exts:
-            yield path
-
-
 
 
 class RadioInterface:
@@ -377,28 +197,96 @@ class UpdaterGUI:
         self.update_thread = None
         self.is_updating = False
         self.selected_version = tk.StringVar(value=VERSION_RELEASE)
+        self.selected_locale = tk.StringVar(value=DEFAULT_LOCALE)
         
         self.setup_ui()
     
     def setup_ui(self):
         """Setup the user interface."""
-        # Title
-        title_frame = ttk.Frame(self.root, padding="10")
+        # Title (dark header)
+        header_bg = "#1f1f1f"
+        header_fg = "#f2f2f2"
+        title_frame = tk.Frame(self.root, bg=header_bg)
         title_frame.pack(fill=tk.X)
-        
-        title_label = ttk.Label(
-            title_frame,
+        title_frame.pack_propagate(False)
+        title_frame.configure(height=90)
+
+        # Layout: text on left, logo on right (if available)
+        text_frame = tk.Frame(title_frame, bg=header_bg)
+        text_frame.pack(side=tk.LEFT, fill=tk.X, expand=True, padx=10, pady=10)
+
+        title_label = tk.Label(
+            text_frame,
             text="Rotorflight Lua Ethos Suite Updater",
-            font=("Arial", 16, "bold")
+            font=("Arial", 16, "bold"),
+            bg=header_bg,
+            fg=header_fg
         )
-        title_label.pack()
-        
-        subtitle_label = ttk.Label(
-            title_frame,
+        title_label.pack(anchor=tk.W)
+
+        subtitle_text = "Open-source Helicopter flight controller"
+
+        # Right side: logo with tagline overlay (logo loads async)
+        logo_frame = tk.Frame(title_frame, bg=header_bg, width=340, height=80)
+        logo_frame.pack(side=tk.RIGHT, padx=(10, 10), pady=10)
+        logo_frame.pack_propagate(False)
+        self.logo_label = None
+        self.logo_image = None
+        subtitle_right = tk.Label(
+            logo_frame,
+            text=subtitle_text,
+            font=("Arial", 9),
+            bg=header_bg,
+            fg=header_fg
+        )
+
+        # Left strapline under title
+        subtitle_left = tk.Label(
+            text_frame,
             text="Update your Ethos radio with the latest suite from GitHub",
-            font=("Arial", 10)
+            font=("Arial", 10),
+            bg=header_bg,
+            fg=header_fg
         )
-        subtitle_label.pack()
+        subtitle_left.pack(anchor=tk.W, pady=(2, 0))
+
+        def set_logo_image(path):
+            try:
+                logo_img = tk.PhotoImage(file=str(path))
+                target_h = 70
+                target_w = 320
+                scale_h = math.ceil(logo_img.height() / target_h)
+                scale_w = math.ceil(logo_img.width() / target_w)
+                scale = max(1, scale_h, scale_w)
+                if scale > 1:
+                    logo_img = logo_img.subsample(scale, scale)
+                self.logo_image = logo_img
+                if not self.logo_label:
+                    self.logo_label = tk.Label(logo_frame, image=self.logo_image, bg=header_bg)
+                else:
+                    self.logo_label.configure(image=self.logo_image)
+
+                # Place logo and overlay tagline within the logo frame
+                self.logo_label.place(relx=1.0, x=0, y=0, anchor=tk.NE)
+                subtitle_right.place(relx=1.0, x=-6, y=52, anchor=tk.NE)
+                subtitle_right.lift()
+            except Exception:
+                pass
+
+        # Async fetch logo to avoid blocking UI
+        def fetch_logo():
+            try:
+                req = Request(LOGO_URL, headers={'User-Agent': 'Mozilla/5.0'})
+                with self.urlopen_insecure(req, timeout=10) as response:
+                    logo_bytes = response.read()
+                tmp_logo = Path(tempfile.gettempdir()) / "rfsuite_logo.png"
+                with open(tmp_logo, "wb") as f:
+                    f.write(logo_bytes)
+                self.root.after(0, lambda: set_logo_image(tmp_logo))
+            except Exception:
+                pass
+
+        threading.Thread(target=fetch_logo, daemon=True).start()
         
         # Version selection frame
         version_frame = ttk.LabelFrame(self.root, text="Version Selection", padding="10")
@@ -432,6 +320,41 @@ class UpdaterGUI:
             variable=self.selected_version,
             value=VERSION_MASTER
         ).pack(side=tk.LEFT, padx=5)
+        # Language selection (within version frame)
+        locale_label = ttk.Label(
+            version_frame,
+            text="Language:",
+            font=("Arial", 9)
+        )
+        locale_label.pack(side=tk.LEFT, padx=10)
+
+        locale_combo = ttk.Combobox(
+            version_frame,
+            textvariable=self.selected_locale,
+            values=AVAILABLE_LOCALES,
+            state="readonly",
+            width=8
+        )
+        locale_combo.pack(side=tk.LEFT, padx=5)
+
+        # Updater update notification (hidden by default)
+        self.update_notice = ttk.Frame(self.root, padding="8")
+        self.update_notice.pack(fill=tk.X, padx=10, pady=(0, 5))
+        self.update_notice.pack_forget()
+
+        self.update_notice_label = ttk.Label(
+            self.update_notice,
+            text="",
+            font=("Arial", 9)
+        )
+        self.update_notice_label.pack(side=tk.LEFT)
+
+        self.update_notice_button = ttk.Button(
+            self.update_notice,
+            text="Open Download Page",
+            command=lambda: webbrowser.open(UPDATER_INFO_URL)
+        )
+        self.update_notice_button.pack(side=tk.RIGHT)
         
         # Status frame
         status_frame = ttk.LabelFrame(self.root, text="Status", padding="10")
@@ -516,6 +439,9 @@ class UpdaterGUI:
             font=("Arial", 8),
             justify=tk.LEFT
         ).pack(anchor=tk.W)
+
+        # Async check for updater updates
+        threading.Thread(target=self.check_updater_update, daemon=True).start()
     
     def log(self, message):
         """Add a message to the log."""
@@ -523,6 +449,44 @@ class UpdaterGUI:
         self.log_text.insert(tk.END, f"[{timestamp}] {message}\n")
         self.log_text.see(tk.END)
         self.root.update_idletasks()
+
+    def urlopen_insecure(self, req, timeout=10):
+        """Open URL without SSL verification (to avoid SSL issues on some systems)."""
+        context = ssl._create_unverified_context()
+        return urlopen(req, timeout=timeout, context=context)
+
+    def _parse_version_tuple(self, version):
+        try:
+            parts = version.strip().split(".")
+            return tuple(int(p) for p in parts)
+        except Exception:
+            return None
+
+    def _is_newer_version(self, current, remote):
+        cur = self._parse_version_tuple(current)
+        rem = self._parse_version_tuple(remote)
+        if cur is None or rem is None:
+            return current != remote
+        return rem > cur
+
+    def check_updater_update(self):
+        """Check if a newer updater version is available."""
+        try:
+            req = Request(UPDATER_RELEASE_JSON_URL, headers={'User-Agent': 'Mozilla/5.0'})
+            with self.urlopen_insecure(req, timeout=10) as response:
+                data = json.loads(response.read().decode())
+            remote_version = data.get("version", "").strip()
+            remote_url = data.get("url") or UPDATER_INFO_URL
+            if remote_version and self._is_newer_version(UPDATER_VERSION, remote_version):
+                msg = f"Updater {remote_version} is available (current {UPDATER_VERSION})."
+                self.root.after(0, lambda: self.show_update_notice(msg, remote_url))
+        except Exception:
+            pass
+
+    def show_update_notice(self, message, url):
+        self.update_notice_label.config(text=message)
+        self.update_notice_button.config(command=lambda: webbrowser.open(url))
+        self.update_notice.pack(fill=tk.X, padx=10, pady=(0, 5))
     
     def set_status(self, message):
         """Update the status label."""
@@ -550,148 +514,7 @@ class UpdaterGUI:
         for root, dirs, files in os.walk(directory):
             total += len(files)
         return total
-
-    def get_deploy_language(self):
-        """Return deploy language from environment or default."""
-        return (os.environ.get("RFSUITE_LANG", "en").strip() or "en").lower()
-
-    def choose_language_for_release(self, staged_out_dir, preferred_lang):
-        """For release/snapshot: if multiple languages exist, choose en when available."""
-        i18n_dir = os.path.join(staged_out_dir, "i18n")
-        if not os.path.isdir(i18n_dir):
-            return preferred_lang
-
-        langs = []
-        for name in os.listdir(i18n_dir):
-            if name.lower().endswith(".json"):
-                langs.append(os.path.splitext(name)[0].lower())
-
-        if len(langs) > 1 and "en" in langs:
-            return "en"
-
-        return preferred_lang
-
-    def file_md5(self, path, chunk=1024 * 1024):
-        import hashlib
-        h = hashlib.md5()
-        with open(path, 'rb', buffering=0) as f:
-            while True:
-                b = f.read(chunk)
-                if not b:
-                    break
-                h.update(b)
-        return h.hexdigest()
-
-    def needs_copy_with_md5(self, srcf, dstf):
-        try:
-            ss = os.stat(srcf)
-        except FileNotFoundError:
-            return False
-        if not os.path.exists(dstf):
-            return True
-        try:
-            ds = os.stat(dstf)
-        except FileNotFoundError:
-            return True
-        if ss.st_size != ds.st_size:
-            return True
-        if (ss.st_mtime - ds.st_mtime) > 2.0:
-            return True
-        try:
-            return self.file_md5(srcf) != self.file_md5(dstf)
-        except Exception:
-            return True
-
-    def copy_tree_update_only(self, src, dest, label="Copying soundpack"):
-        """Copy directory tree, updating only changed files."""
-        if not os.path.isdir(src):
-            self.log(f"  ⚠ Source not found: {src}")
-            return False
-
-        files = []
-        for r, _, fs in os.walk(src):
-            for f in fs:
-                s = os.path.join(r, f)
-                rel = os.path.relpath(s, src)
-                d = os.path.join(dest, rel)
-                files.append((s, d, rel))
-
-        if not files:
-            self.log("  No files to copy.")
-            return True
-
-        self.set_progress_mode('determinate', maximum=len(files))
-        copied = 0
-        for s, d, rel in files:
-            if not self.is_updating:
-                return False
-            os.makedirs(os.path.dirname(d), exist_ok=True)
-            if self.needs_copy_with_md5(s, d):
-                try:
-                    shutil.copy2(s, d)
-                except Exception as e:
-                    self.log(f"  ⚠ Failed to copy {rel}: {e}")
-            copied += 1
-            if copied % 25 == 0 or copied == len(files):
-                percent = (copied / len(files)) * 100
-                self.update_progress(copied, f"{label}: {copied}/{len(files)} ({percent:.1f}%)")
-        return True
-
-    def resolve_i18n_tags_in_tree(self, json_path, root_dir):
-        """Resolve @i18n(...)@ tags in-place under root_dir using json_path."""
-        import json
-        from pathlib import Path
-
-        translations = None
-        try:
-            with open(json_path, 'r', encoding='utf-8') as f:
-                translations = json.load(f)
-        except Exception as e:
-            self.log(f"  ⚠ Failed to load i18n JSON: {e}")
-            return False
-
-        root = Path(root_dir)
-        total_files_changed = 0
-        total_replacements = 0
-        unresolved_agg = {}
-
-        files = list(_iter_source_files(root))
-        if files:
-            self.set_progress_mode('determinate', maximum=len(files))
-        for idx, file_path in enumerate(files, 1):
-            if not self.is_updating:
-                return False
-            try:
-                before = file_path.read_text(encoding='utf-8')
-            except Exception:
-                continue
-
-            stats = {}
-            new_text, count = _replace_tags_in_text(before, translations, stats)
-
-            if count:
-                try:
-                    file_path.write_text(new_text, encoding='utf-8')
-                    total_files_changed += 1
-                    total_replacements += count
-                except Exception as e:
-                    self.log(f"  ⚠ Failed to write {file_path}: {e}")
-
-            unresolved = stats.get('unresolved', {})
-            for k, c in unresolved.items():
-                unresolved_agg[k] = unresolved_agg.get(k, 0) + c
-
-            if idx % 25 == 0 or idx == len(files):
-                percent = (idx / len(files)) * 100 if files else 0
-                self.update_progress(idx, f"Resolving i18n: {idx}/{len(files)} ({percent:.1f}%)")
-
-        self.log(f"✓ i18n resolved — files changed: {total_files_changed}, total replacements: {total_replacements}")
-        if unresolved_agg:
-            self.log("  Unresolved keys (top 10):")
-            for key, count in sorted(unresolved_agg.items(), key=lambda kv: (-kv[1], kv[0]))[:10]:
-                self.log(f"    {key}: {count} occurrence(s)")
-        return True
-
+    
     def copy_tree_with_progress(self, src, dst):
         """Copy directory tree with progress updates."""
         # Count total files
@@ -719,6 +542,7 @@ class UpdaterGUI:
                 try:
                     shutil.copy2(src_file, dst_file)
                     copied += 1
+                    time.sleep(COPY_SETTLE_SECONDS)
                     
                     # Update progress
                     percent = (copied / total_files) * 100
@@ -791,7 +615,7 @@ class UpdaterGUI:
         
         return True
     
-    def get_download_url_and_name(self):
+    def get_download_url_and_name(self, locale):
         """Get the download URL and version name based on selected version."""
         import json
         
@@ -801,47 +625,73 @@ class UpdaterGUI:
             # Master branch
             url = f"{GITHUB_REPO_URL}/archive/refs/heads/master.zip"
             name = "master"
-            return url, name
+            return url, name, False
         
         elif version_type == VERSION_SNAPSHOT:
             # Snapshot branch - try to get pre-release from GitHub API
             try:
                 self.log("Fetching latest pre-release information...")
                 req = Request(f"{GITHUB_API_URL}/releases", headers={'User-Agent': 'Mozilla/5.0'})
-                with urlopen(req, timeout=10) as response:
+                with self.urlopen_insecure(req, timeout=DOWNLOAD_TIMEOUT) as response:
                     data = json.loads(response.read().decode())
                     # Find first pre-release
                     for release in data:
                         if release.get('prerelease', False):
                             tag_name = release.get('tag_name', '')
                             if tag_name:
+                                version = tag_name.split("/", 1)[1] if "/" in tag_name else tag_name
+                                asset_name = f"rotorflight-lua-ethos-suite-{version}-{locale}.zip"
+                                for asset in release.get("assets", []):
+                                    if asset.get("name") == asset_name:
+                                        self.log(f"✓ Found latest pre-release asset: {asset_name}")
+                                        return asset.get("browser_download_url"), tag_name, True
+                                if locale != DEFAULT_LOCALE:
+                                    fallback_name = f"rotorflight-lua-ethos-suite-{version}-{DEFAULT_LOCALE}.zip"
+                                    for asset in release.get("assets", []):
+                                        if asset.get("name") == fallback_name:
+                                            self.log(f"⚠ Locale '{locale}' asset not found; using {DEFAULT_LOCALE}")
+                                            return asset.get("browser_download_url"), tag_name, True
+                                # Fall back to source zip if asset missing
                                 url = f"{GITHUB_REPO_URL}/archive/refs/tags/{tag_name}.zip"
-                                self.log(f"✓ Found latest pre-release: {tag_name}")
-                                return url, tag_name
+                                self.log(f"✓ Found latest pre-release tag: {tag_name} (no asset)")
+                                return url, tag_name, False
                     # No pre-release found, fall back to master
                     self.log("⚠ No pre-release found, using master branch")
                     url = f"{GITHUB_REPO_URL}/archive/refs/heads/master.zip"
                     name = "master (no snapshot)"
-                    return url, name
+                    return url, name, False
             except Exception as e:
                 self.log(f"⚠ Failed to fetch pre-release info: {e}")
                 self.log("  Falling back to master branch")
                 url = f"{GITHUB_REPO_URL}/archive/refs/heads/master.zip"
                 name = "master (fallback)"
-                return url, name
+                return url, name, False
         
         elif version_type == VERSION_RELEASE:
             # Latest release
             try:
                 self.log("Fetching latest release information...")
                 req = Request(f"{GITHUB_API_URL}/releases/latest", headers={'User-Agent': 'Mozilla/5.0'})
-                with urlopen(req, timeout=10) as response:
+                with self.urlopen_insecure(req, timeout=DOWNLOAD_TIMEOUT) as response:
                     data = json.loads(response.read().decode())
                     tag_name = data.get('tag_name', '')
                     if tag_name:
+                        version = tag_name.split("/", 1)[1] if "/" in tag_name else tag_name
+                        asset_name = f"rotorflight-lua-ethos-suite-{version}-{locale}.zip"
+                        for asset in data.get("assets", []):
+                            if asset.get("name") == asset_name:
+                                self.log(f"✓ Found latest release asset: {asset_name}")
+                                return asset.get("browser_download_url"), tag_name, True
+                        if locale != DEFAULT_LOCALE:
+                            fallback_name = f"rotorflight-lua-ethos-suite-{version}-{DEFAULT_LOCALE}.zip"
+                            for asset in data.get("assets", []):
+                                if asset.get("name") == fallback_name:
+                                    self.log(f"⚠ Locale '{locale}' asset not found; using {DEFAULT_LOCALE}")
+                                    return asset.get("browser_download_url"), tag_name, True
+                        # Fall back to source zip if asset missing
                         url = f"{GITHUB_REPO_URL}/archive/refs/tags/{tag_name}.zip"
-                        self.log(f"✓ Found latest release: {tag_name}")
-                        return url, tag_name
+                        self.log(f"✓ Found latest release tag: {tag_name} (no asset)")
+                        return url, tag_name, False
                     else:
                         raise RuntimeError("No tag_name in release data")
             except Exception as e:
@@ -849,12 +699,196 @@ class UpdaterGUI:
                 self.log("  Falling back to master branch")
                 url = f"{GITHUB_REPO_URL}/archive/refs/heads/master.zip"
                 name = "master (fallback)"
-                return url, name
+                return url, name, False
         
         # Default fallback
         url = f"{GITHUB_REPO_URL}/archive/refs/heads/master.zip"
         name = "master"
-        return url, name
+        return url, name, False
+
+    def is_git_available(self):
+        """Check if git is available."""
+        try:
+            result = subprocess.run(
+                ["git", "--version"],
+                capture_output=True,
+                text=True,
+                timeout=5,
+                creationflags=subprocess.CREATE_NO_WINDOW if sys.platform == "win32" else 0
+            )
+            return result.returncode == 0
+        except Exception:
+            return False
+
+    def sparse_checkout_master(self, dest_dir, locale):
+        """Sparse checkout required folders from master into dest_dir."""
+        if not self.is_git_available():
+            self.log("⚠ Git not available; falling back to ZIP download")
+            return False
+
+        self.log("Using git sparse checkout for master...")
+        os.makedirs(dest_dir, exist_ok=True)
+
+        def run_git(args, timeout=60):
+            cmd = ["git"] + args
+            self.log(f"  Git: {' '.join(cmd)}")
+            result = subprocess.run(
+                cmd,
+                cwd=dest_dir,
+                capture_output=True,
+                text=True,
+                timeout=timeout,
+                creationflags=subprocess.CREATE_NO_WINDOW if sys.platform == "win32" else 0
+            )
+            if result.stdout:
+                for line in result.stdout.strip().splitlines():
+                    if line.strip():
+                        self.log(f"    [git] {line}")
+            if result.stderr:
+                for line in result.stderr.strip().splitlines():
+                    if line.strip():
+                        self.log(f"    [git] {line}")
+            return result
+
+        # Initialize repository and configure sparse checkout
+        init = run_git(["init"])
+        if init.returncode != 0:
+            self.log(f"⚠ Git init failed: {init.stderr.strip()}")
+            return False
+
+        run_git(["remote", "add", "origin", GITHUB_REPO_URL + ".git"])
+        run_git(["config", "core.sparseCheckout", "true"])
+
+        sparse_file = os.path.join(dest_dir, ".git", "info", "sparse-checkout")
+        with open(sparse_file, "w", encoding="utf-8") as f:
+            f.write("src/rfsuite/\n")
+            f.write(".vscode/scripts/\n")
+            # Grab all soundpacks to avoid missing locale audio during master installs
+            f.write("bin/sound-generator/soundpack/\n")
+
+        fetch = run_git(["fetch", "--depth", "1", "--progress", "origin", "master"], timeout=180)
+        if fetch.returncode != 0:
+            self.log(f"⚠ Git fetch failed: {fetch.stderr.strip()}")
+            return False
+
+        checkout = run_git(["checkout", "FETCH_HEAD"])
+        if checkout.returncode != 0:
+            self.log(f"⚠ Git checkout failed: {checkout.stderr.strip()}")
+            return False
+
+        self.log("✓ Sparse checkout completed")
+        return True
+
+    def copy_sound_pack(self, repo_dir, dest_dir, locale):
+        """Copy sound pack for selected locale into destination tree."""
+        locale = locale or DEFAULT_LOCALE
+        src = os.path.join(repo_dir, "bin", "sound-generator", "soundpack", locale)
+        if not os.path.isdir(src):
+            if locale != DEFAULT_LOCALE:
+                self.log(f"⚠ Sound pack for '{locale}' not found; using {DEFAULT_LOCALE}")
+            locale = DEFAULT_LOCALE
+            src = os.path.join(repo_dir, "bin", "sound-generator", "soundpack", locale)
+        if not os.path.isdir(src):
+            self.log("⚠ Sound pack not found; skipping audio copy")
+            return False
+
+        dest = os.path.join(dest_dir, "audio", locale)
+        try:
+            if os.path.isdir(dest):
+                shutil.rmtree(dest)
+            os.makedirs(dest, exist_ok=True)
+            shutil.copytree(src, dest, dirs_exist_ok=True)
+            self.log(f"✓ Audio pack copied: {locale}")
+            return True
+        except Exception as e:
+            self.log(f"⚠ Failed to copy audio pack: {e}")
+            return False
+
+    def locate_source_dir(self, extract_dir):
+        """Locate the rfsuite source directory in extracted content."""
+        possible_paths = [
+            os.path.join(extract_dir, "scripts", TARGET_NAME),  # prebuilt asset layout
+            os.path.join(extract_dir, "src", TARGET_NAME),      # repo layout
+            os.path.join(extract_dir, TARGET_NAME),             # direct
+        ]
+        for path in possible_paths:
+            if os.path.isdir(path):
+                return path
+        return None
+
+    def get_master_commit_suffix(self):
+        """Fetch the latest master commit SHA and return commit-<sha7>."""
+        import json
+        try:
+            req = Request(f"{GITHUB_API_URL}/commits/master", headers={'User-Agent': 'Mozilla/5.0'})
+            with self.urlopen_insecure(req, timeout=DOWNLOAD_TIMEOUT) as response:
+                data = json.loads(response.read().decode())
+                sha = data.get("sha", "")
+                if sha:
+                    return f"commit-{sha[:7]}"
+        except Exception as e:
+            self.log(f"⚠ Failed to fetch master commit SHA: {e}")
+        return "master"
+
+    def derive_version_suffix(self, version_type, version_name):
+        """Derive the version suffix to write into main.lua based on workflows."""
+        # Strip common tag prefixes
+        if version_name.startswith("release/"):
+            return version_name.split("/", 1)[1]
+        if version_name.startswith("snapshot/"):
+            return version_name.split("/", 1)[1]
+        if version_type == VERSION_MASTER:
+            return self.get_master_commit_suffix()
+        # Fallback: use a sanitized first token
+        return (version_name.split()[0] if version_name else "master")
+
+    def update_main_lua_version(self, main_lua_path, version_suffix):
+        """Update the version suffix in main.lua."""
+        try:
+            with open(main_lua_path, "r", encoding="utf-8") as f:
+                content = f.read()
+        except Exception as e:
+            self.log(f"⚠ Unable to read main.lua for version update: {e}")
+            return False
+
+        pattern = re.compile(r'(version\s*=\s*\{[^}]*suffix\s*=\s*")([^"]*)(")', re.S)
+        if not pattern.search(content):
+            self.log("⚠ Version suffix pattern not found in main.lua")
+            return False
+
+        updated = pattern.sub(lambda m: f"{m.group(1)}{version_suffix}{m.group(3)}", content)
+        try:
+            with open(main_lua_path, "w", encoding="utf-8") as f:
+                f.write(updated)
+            self.log(f"✓ Updated main.lua version suffix to '{version_suffix}'")
+            return True
+        except Exception as e:
+            self.log(f"⚠ Unable to write main.lua version update: {e}")
+            return False
+
+    def read_main_lua_version(self, main_lua_path):
+        """Read the full version string from main.lua."""
+        try:
+            with open(main_lua_path, "r", encoding="utf-8") as f:
+                content = f.read()
+        except Exception as e:
+            self.log(f"⚠ Unable to read main.lua for version info: {e}")
+            return None
+
+        pattern = re.compile(
+            r'version\s*=\s*\{[^}]*major\s*=\s*(\d+)[^}]*minor\s*=\s*(\d+)[^}]*revision\s*=\s*(\d+)[^}]*suffix\s*=\s*"([^"]+)"',
+            re.S
+        )
+        match = pattern.search(content)
+        if not match:
+            self.log("⚠ Could not parse version fields from main.lua")
+            return None
+
+        major, minor, revision, suffix = match.groups()
+        prefix = f"{major}.{minor}.{revision}-"
+        if suffix.startswith(prefix):
+            return suffix
+        return f"{major}.{minor}.{revision}-{suffix}"
     
     def start_update(self):
         """Start the update process in a background thread."""
@@ -960,158 +994,154 @@ class UpdaterGUI:
                 # Radio was already mounted, scripts_dir is already set
                 self.log("Skipping mount wait (radio already mounted)")
             
-            # Step 5: Download suite from GitHub
-            self.set_status("Downloading suite from GitHub...")
+            if not self.is_updating:
+                return
             
-            # Get download URL based on selected version
-            download_url, version_name = self.get_download_url_and_name()
-            self.log(f"Selected version: {version_name}")
-            self.log(f"Downloading from: {download_url}")
+            # Step 5: Download suite from GitHub (or git sparse checkout for master)
+            self.set_status("Preparing download...")
+
+            version_type = self.selected_version.get()
+            version_name = "master" if version_type == VERSION_MASTER else ""
+            version_suffix = self.derive_version_suffix(version_type, version_name)
+            locale = self.selected_locale.get() or DEFAULT_LOCALE
+            self.log(f"Selected version: {version_name or version_type}")
+            self.log(f"Selected locale: {locale}")
+            self.log(f"Version suffix for main.lua: {version_suffix}")
+            is_asset = False
             
             temp_dir = tempfile.mkdtemp(prefix="rfsuite-update-")
             # Sanitize version name for filename (replace / with -)
             safe_version_name = version_name.replace('/', '-').replace('\\', '-')
             zip_path = os.path.join(temp_dir, f"{safe_version_name}.zip")
-            
-            try:
-                req = Request(download_url, headers={'User-Agent': 'Mozilla/5.0'})
-                with urlopen(req, timeout=30) as response:
-                    total_size = int(response.headers.get('content-length', 0))
-                    downloaded = 0
-                    
-                    with open(zip_path, 'wb') as f:
-                        while True:
-                            if not self.is_updating:
-                                return
-                            
-                            chunk = response.read(8192)
-                            if not chunk:
-                                break
-                            
-                            f.write(chunk)
-                            downloaded += len(chunk)
-                            
-                            if total_size > 0:
-                                percent = (downloaded / total_size) * 100
-                                self.log(f"  Downloaded: {downloaded}/{total_size} bytes ({percent:.1f}%)")
-                
-                self.log(f"✓ Downloaded {downloaded} bytes")
-            except (URLError, HTTPError) as e:
-                self.log(f"✗ Download failed: {e}")
-                raise
+
+            # For master, prefer sparse checkout to avoid full repo download
+            repo_dir = None
+            if version_type == VERSION_MASTER:
+                self.set_status("Fetching master via git...")
+                self.set_progress_mode('indeterminate')
+                self.log("Git sparse checkout: src/rfsuite/, .vscode/scripts/, bin/sound-generator/soundpack/")
+                repo_dir = os.path.join(temp_dir, "repo")
+                if not self.sparse_checkout_master(repo_dir, locale):
+                    repo_dir = None
+                else:
+                    self.log("✓ Using sparse checkout; skipping ZIP download")
+
+            if repo_dir is None:
+                # Get download URL based on selected version (release/snapshot or master fallback)
+                download_url, version_name, is_asset = self.get_download_url_and_name(locale)
+                if not download_url:
+                    raise RuntimeError("No download URL available")
+                if version_type != VERSION_MASTER:
+                    version_suffix = self.derive_version_suffix(version_type, version_name)
+                    self.log(f"Selected version: {version_name}")
+                    self.log(f"Version suffix for main.lua: {version_suffix}")
+                self.log(f"Downloading from: {download_url}")
+                try:
+                    req = Request(download_url, headers={'User-Agent': 'Mozilla/5.0'})
+                    attempt = 0
+                    while True:
+                        attempt += 1
+                        try:
+                            self.log(f"  Download attempt {attempt}/{DOWNLOAD_RETRIES} (timeout {DOWNLOAD_TIMEOUT}s)")
+                            with self.urlopen_insecure(req, timeout=DOWNLOAD_TIMEOUT) as response:
+                                total_size = int(response.headers.get('content-length', 0))
+                                size_known = total_size > 0
+                                downloaded = 0
+
+                                if size_known:
+                                    self.set_progress_mode('determinate', maximum=total_size)
+                                else:
+                                    total_size = 50 * 1024 * 1024  # 50MB estimate
+                                    self.set_progress_mode('determinate', maximum=total_size)
+                                    self.log("  Download size unknown (no content-length); estimating 50MB")
+
+                                last_log_percent = -1
+
+                                with open(zip_path, 'wb') as f:
+                                    while True:
+                                        if not self.is_updating:
+                                            return
+
+                                        chunk = response.read(8192)
+                                        if not chunk:
+                                            break
+
+                                        f.write(chunk)
+                                        downloaded += len(chunk)
+
+                                        percent = (downloaded / total_size) * 100 if total_size > 0 else 0
+                                        if int(percent) != last_log_percent:
+                                            last_log_percent = int(percent)
+                                            if size_known:
+                                                self.log(f"  Downloaded: {downloaded}/{total_size} bytes ({percent:.1f}%)")
+                                            else:
+                                                self.log(f"  Downloaded: {downloaded}/{total_size} bytes ({percent:.1f}%) (estimated)")
+                                        self.update_progress(downloaded, f"Downloading... {percent:.1f}%")
+
+                            break
+                        except (URLError, HTTPError) as e:
+                            if attempt >= DOWNLOAD_RETRIES:
+                                raise
+                            self.log(f"  Download failed: {e}. Retrying in {DOWNLOAD_RETRY_DELAY}s...")
+                            time.sleep(DOWNLOAD_RETRY_DELAY)
+                    self.log(f"✓ Downloaded {downloaded} bytes")
+                except (URLError, HTTPError) as e:
+                    self.log(f"✗ Download failed: {e}")
+                    raise
             
             if not self.is_updating:
                 return
             
-            # Step 6: Extract archive
-            self.set_status("Extracting archive...")
-            self.log("Extracting downloaded archive...")
-            
-            extract_dir = os.path.join(temp_dir, "extracted")
-            
-            try:
-                with zipfile.ZipFile(zip_path, 'r') as zip_ref:
-                    zip_ref.extractall(extract_dir)
-                self.log("✓ Archive extracted")
-            except Exception as e:
-                self.log(f"✗ Extraction failed: {e}")
-                raise
+            # Step 6: Extract archive (if we downloaded a zip)
+            extract_dir = None
+            if repo_dir is None:
+                self.set_status("Extracting archive...")
+                self.log("Extracting downloaded archive...")
+
+                extract_dir = os.path.join(temp_dir, "extracted")
+
+                try:
+                    with zipfile.ZipFile(zip_path, 'r') as zip_ref:
+                        zip_ref.extractall(extract_dir)
+                    self.log("✓ Archive extracted")
+                except Exception as e:
+                    self.log(f"✗ Extraction failed: {e}")
+                    raise
             
             if not self.is_updating:
                 return
             
             # Step 7: Find source directory
             self.log("Locating source files...")
-            
-            # GitHub creates a folder like "rotorflight-lua-ethos-suite-master"
-            extracted_items = os.listdir(extract_dir)
-            if not extracted_items:
-                raise RuntimeError("Extracted archive is empty")
-            
-            repo_dir = os.path.join(extract_dir, extracted_items[0])
-            
-            # Try different possible locations for the source
-            possible_paths = [
-                os.path.join(repo_dir, "src", TARGET_NAME),      # Standard: src/rfsuite
-                os.path.join(repo_dir, "scripts", TARGET_NAME),  # Alternative: scripts/rfsuite
-                os.path.join(repo_dir, TARGET_NAME),             # Direct: rfsuite
-            ]
-            
-            src_dir = None
-            for path in possible_paths:
-                if os.path.isdir(path):
-                    src_dir = path
-                    self.log(f"✓ Found source: {os.path.relpath(path, extract_dir)}")
-                    break
+
+            if repo_dir is None:
+                # GitHub creates a folder like "rotorflight-lua-ethos-suite-master"
+                extracted_items = os.listdir(extract_dir)
+                if not extracted_items:
+                    raise RuntimeError("Extracted archive is empty")
+                repo_dir = os.path.join(extract_dir, extracted_items[0])
+
+            src_dir = self.locate_source_dir(repo_dir)
+            if src_dir:
+                self.log(f"✓ Found source: {os.path.relpath(src_dir, repo_dir)}")
             
             if not src_dir:
                 self.log(f"✗ Source directory not found in any of these locations:")
-                for path in possible_paths:
-                    self.log(f"  - {os.path.relpath(path, extract_dir)}")
+                for path in [
+                    os.path.join(repo_dir, "scripts", TARGET_NAME),
+                    os.path.join(repo_dir, "src", TARGET_NAME),
+                    os.path.join(repo_dir, TARGET_NAME),
+                ]:
+                    self.log(f"  - {os.path.relpath(path, repo_dir)}")
                 raise RuntimeError(f"Could not find {TARGET_NAME} in extracted archive")
             
             if not self.is_updating:
                 return
             
-            # Step 8: Stage files locally (apply steps before radio copy)
-            lang = self.get_deploy_language()
-            self.set_status("Staging files...")
-            self.log(f"Staging files for language: {lang}")
-
-            stage_dir = os.path.join(temp_dir, "stage")
-            staged_out_dir = os.path.join(stage_dir, TARGET_NAME)
-
-            if os.path.isdir(staged_out_dir):
-                self.remove_tree_with_progress(staged_out_dir)
-
-            if not self.copy_tree_with_progress(src_dir, staged_out_dir):
-                self.log("⚠ Staging cancelled")
-                return
-
-            if not self.is_updating:
-                return
-
-            # Step 9: Run deploy steps in-process (i18n + soundpack)
-            self.set_status("Applying deploy steps...")
-            self.log("Applying i18n and soundpack steps...")
-
-            if self.selected_version.get() in (VERSION_RELEASE, VERSION_SNAPSHOT):
-                chosen_lang = self.choose_language_for_release(staged_out_dir, lang)
-                if chosen_lang != lang:
-                    self.log(f"  Multiple languages detected; using {chosen_lang}")
-                    lang = chosen_lang
-
-            i18n_json = None
-            for base_path in [
-                os.path.join(staged_out_dir, "i18n", f"{lang}.json"),
-                os.path.join(repo_dir, "src", TARGET_NAME, "i18n", f"{lang}.json"),
-                os.path.join(repo_dir, "scripts", TARGET_NAME, "i18n", f"{lang}.json"),
-            ]:
-                if os.path.isfile(base_path):
-                    i18n_json = base_path
-                    break
-
-            if i18n_json:
-                self.log(f"  Resolving i18n tags using {os.path.basename(i18n_json)}")
-                self.resolve_i18n_tags_in_tree(i18n_json, staged_out_dir)
-            else:
-                self.log("⚠ i18n JSON not found; skipping i18n resolution")
-
-            soundpack_src = os.path.join(repo_dir, "bin", "sound-generator", "soundpack", lang)
-            soundpack_dst = os.path.join(staged_out_dir, "audio", lang)
-            if os.path.isdir(soundpack_src):
-                self.log(f"  Copying soundpack: {lang}")
-                self.copy_tree_update_only(soundpack_src, soundpack_dst, label="Soundpack")
-            else:
-                self.log(f"⚠ Soundpack not found at {soundpack_src}; skipping")
-
-            if not self.is_updating:
-                return
-
-            # Step 10: Copy files to radio
+            # Step 8: Copy files to radio
             dest_dir = os.path.join(scripts_dir, TARGET_NAME)
             self.set_status("Copying files to radio...")
-            self.log("Copying staged files to radio...")
+            self.log("Copying new files to radio...")
             
             # Remove old installation
             if os.path.isdir(dest_dir):
@@ -1125,15 +1155,88 @@ class UpdaterGUI:
             
             # Copy new files
             self.log("  Copying new files...")
-            if not self.copy_tree_with_progress(staged_out_dir, dest_dir):
+            if not self.copy_tree_with_progress(src_dir, dest_dir):
                 self.log("⚠ Copy cancelled")
                 return
             
             self.log(f"✓ Files copied to radio successfully")
+
+            # Ensure audio pack matches selected locale for master/zip builds
+            if not is_asset:
+                self.copy_sound_pack(repo_dir, dest_dir, locale)
+
+            # Update main.lua version suffix only for master (release/snapshot assets already stamped)
+            main_lua_path = os.path.join(dest_dir, "main.lua")
+            if version_type == VERSION_MASTER:
+                if os.path.isfile(main_lua_path):
+                    self.update_main_lua_version(main_lua_path, version_suffix)
+                else:
+                    self.log(f"⚠ main.lua not found at {main_lua_path} for version update")
             
             if not self.is_updating:
                 return
-
+            
+            # Step 9: Compile i18n translations (skip for prebuilt assets)
+            if not is_asset:
+                self.set_status("Compiling translations...")
+                self.log("Compiling i18n translations...")
+                self.set_progress_mode('indeterminate')
+                
+                try:
+                    # Find i18n JSON file in the extracted repo (try multiple locations)
+                    i18n_json = None
+                    locale = self.selected_locale.get() or DEFAULT_LOCALE
+                    for base_path in [
+                        os.path.join(repo_dir, "src", TARGET_NAME, "i18n", f"{locale}.json"),
+                        os.path.join(repo_dir, "scripts", TARGET_NAME, "i18n", f"{locale}.json"),
+                        os.path.join(repo_dir, TARGET_NAME, "i18n", f"{locale}.json"),
+                        os.path.join(repo_dir, "src", TARGET_NAME, "i18n", f"{DEFAULT_LOCALE}.json"),
+                        os.path.join(repo_dir, "scripts", TARGET_NAME, "i18n", f"{DEFAULT_LOCALE}.json"),
+                        os.path.join(repo_dir, TARGET_NAME, "i18n", f"{DEFAULT_LOCALE}.json"),
+                    ]:
+                        if os.path.isfile(base_path):
+                            i18n_json = base_path
+                            break
+                    
+                    resolver_script = os.path.join(repo_dir, ".vscode", "scripts", "resolve_i18n_tags.py")
+                    
+                    if i18n_json and os.path.isfile(resolver_script):
+                        self.log(f"  Using i18n JSON: {os.path.basename(i18n_json)}")
+                        self.log(f"  Running resolver script...")
+                        
+                        result = subprocess.run(
+                            [sys.executable, resolver_script, "--json", i18n_json, "--root", dest_dir],
+                            capture_output=True,
+                            text=True,
+                            timeout=60
+                        )
+                        
+                        if result.returncode == 0:
+                            self.log("✓ i18n translations compiled successfully")
+                            if result.stdout:
+                                for line in result.stdout.strip().split('\n'):
+                                    if line.strip():
+                                        self.log(f"  {line}")
+                        else:
+                            self.log(f"⚠ i18n compilation failed (exit code {result.returncode})")
+                            if result.stderr:
+                                for line in result.stderr.strip().split('\n')[:5]:  # Show first 5 error lines
+                                    if line.strip():
+                                        self.log(f"  {line}")
+                    else:
+                        self.log("⚠ i18n files not found, skipping translation compilation")
+                        if not os.path.isfile(i18n_json):
+                            self.log(f"  Missing: {i18n_json}")
+                        if not os.path.isfile(resolver_script):
+                            self.log(f"  Missing: {resolver_script}")
+                except subprocess.TimeoutExpired:
+                    self.log("⚠ i18n compilation timed out")
+                except Exception as e:
+                    self.log(f"⚠ i18n compilation error: {e}")
+            
+            if not self.is_updating:
+                return
+            
             # Step 11: Cleanup
             self.set_status("Cleaning up...")
             self.log("Cleaning up temporary files...")
@@ -1152,6 +1255,11 @@ class UpdaterGUI:
             self.log("✓ UPDATE COMPLETED SUCCESSFULLY!")
             self.log("=" * 50)
             self.log("")
+            full_version = self.read_main_lua_version(main_lua_path) if 'main_lua_path' in locals() else None
+            if full_version:
+                self.log(f"Installed version: {full_version}")
+            else:
+                self.log(f"Installed version suffix: {version_suffix}")
             self.log("You can now disconnect your radio and restart it.")
             self.log("The new Rotorflight Lua Ethos Suite is ready to use.")
             
@@ -1216,12 +1324,29 @@ def check_dependencies():
 
 def main():
     """Main entry point."""
-    if not check_dependencies():
+    try:
+        if not check_dependencies():
+            sys.exit(1)
+        
+        root = tk.Tk()
+        app = UpdaterGUI(root)
+        root.mainloop()
+    except Exception:
+        error_log = Path(__file__).resolve().parent / "updater_error.log"
+        with open(error_log, "w", encoding="utf-8") as f:
+            f.write(traceback.format_exc())
+        try:
+            if 'tk' in sys.modules:
+                root = tk.Tk()
+                root.withdraw()
+                messagebox.showerror(
+                    "Updater Error",
+                    f"Updater failed to start.\n\nDetails written to:\n{error_log}"
+                )
+                root.destroy()
+        except Exception:
+            pass
         sys.exit(1)
-    
-    root = tk.Tk()
-    app = UpdaterGUI(root)
-    root.mainloop()
 
 
 if __name__ == "__main__":
